feat: add class of draw pages
