+++ package.json
@@ -4,44 +4,44 @@
   "description": "",
   "main": "webpack.base.config.js",
   "scripts": {
-      "start": "webpack serve --open --config ./webpack.config.js --env mode=dev",
-      "build": "webpack --config ./webpack.config.js --env mode=prod",
-      "lint": "eslint . --ext .ts"
+    "start": "webpack serve --open --config ./webpack.config.js --env mode=dev",
+    "build": "webpack --config ./webpack.config.js --env mode=prod",
+    "lint": "eslint . --ext .ts"
   },
   "repository": {
-      "type": "git",
-      "url": ""
+    "type": "git",
+    "url": ""
   },
   "keywords": [],
   "author": "Ilya Kotsur",
   "license": "ISC",
   "homepage": "",
   "devDependencies": {
-      "@types/wnumb": "^1.2.1",
-      "@typescript-eslint/eslint-plugin": "^5.30.0",
-      "@typescript-eslint/parser": "^5.30.0",
-      "clean-webpack-plugin": "^3.0.0",
-      "copy-webpack-plugin": "^7.0.0",
-      "css-loader": "^5.1.0",
-      "eslint": "^7.27.0",
-      "eslint-config-prettier": "^8.3.0",
-      "eslint-plugin-import": "^2.23.3",
-      "eslint-plugin-prettier": "^3.4.0",
-      "file-loader": "^6.2.0",
-      "html-webpack-plugin": "^5.2.0",
-      "mini-css-extract-plugin": "^2.6.1",
-      "nouislider": "^15.6.1",
-      "prettier": "2.2.1",
-      "sass": "^1.53.0",
-      "sass-loader": "^13.0.2",
-      "simplebar": "^5.3.8",
-      "style-loader": "^2.0.0",
-      "ts-loader": "^9.3.1",
-      "typescript": "^4.7.4",
-      "webpack": "^5.73.0",
-      "webpack-cli": "^4.5.0",
-      "webpack-dev-server": "^3.11.2",
-      "webpack-merge": "^5.7.3",
-      "wnumb": "^1.2.0"
+    "@types/wnumb": "^1.2.1",
+    "@typescript-eslint/eslint-plugin": "^5.30.0",
+    "@typescript-eslint/parser": "^5.30.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^7.0.0",
+    "css-loader": "^5.1.0",
+    "eslint": "^7.27.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.23.3",
+    "eslint-plugin-prettier": "^3.4.0",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.2.0",
+    "mini-css-extract-plugin": "^2.6.1",
+    "nouislider": "^15.6.1",
+    "prettier": "2.2.1",
+    "sass": "^1.53.0",
+    "sass-loader": "^13.0.2",
+    "simplebar": "^5.3.8",
+    "style-loader": "^2.0.0",
+    "ts-loader": "^9.3.1",
+    "typescript": "^4.7.4",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.5.0",
+    "webpack-dev-server": "^3.11.2",
+    "webpack-merge": "^5.7.3",
+    "wnumb": "^1.2.0"
   }
 }
