+++ src/index.html
@@ -13,95 +13,5 @@
     />
     <title>Online store</title>
   </head>
-  <body>
-    <header class="header">
-      <div class="header__container container">
-        <div class="container__logo logo">
-          <div class="logo__pict">&nbsp;</div>
-          <h1 class="logo__caption"><a class="logo__home" href="./index.html">Online Store</a></h1>
-        </div>
-        <div class="container__price-basket price-basket">
-          <p class="price-basket__name">Cart total: <span class="price-basket__name_count">0</span>$</p>
-        </div>
-        <div class="container__to-basket to-basket">
-          <div class="to-basket__pict">&nbsp;</div>
-        </div>
-      </div>
-    </header>
-    <main class="main">
-      <div class="container">
-        <div class="main__wrapper">
-          <div class="main__aside aside">
-            <div class="aside__filters filters">
-            </div>
-          </div>
-          <div class="main__catalog catalog">
-            <div class="catalog__panel">
-              <div class="catalog__search search">
-                <form class="search__form">
-                  <input maxlength="15" type="search" placeholder="Search product" value="" class="search__input">
-                </form>
-              </div>
-              <div class="catalog__copy copy">
-                <button class="copy__btn">Copy Link</button>
-              </div>
-              <div class="catalog__sort sort">
-                <button class="sort__button">
-                  <p class="sort__text">Choose Sort</p>
-                </button>
-              </div>
-              <div class="catalog__typeView typeView">
-                <button class="typeView__icon">≡</button>
-              </div>
-              <div class="sort__chooseList chooseList">
-                <form class="chooseList__form">
-                  <div class="chooseList__typeSort">
-                    <input type="radio" id="popularityUp" name="sortList">
-                    <label for="popularityUp">By popularity(Ascending)</label>
-                  </div>
-                  <div class="chooseList__typeSort">
-                    <input type="radio" id="popularityLow" name="sortList">
-                    <label for="popularityLow">By popularity(Descending)</label>
-                  </div>
-                  <div class="chooseList__typeSort">
-                    <input type="radio" id="priceUp" name="sortList">
-                    <label for="priceUp">By price(Ascending)</label>
-                  </div>
-                  <div class="chooseList__typeSort">
-                    <input type="radio" id="priceLow" name="sortList">
-                    <label for="priceLow">By price(Descending)</label>
-                  </div>
-                </form>
-              </div>
-            </div>
-            <div class="catalog__products"></div>
-          </div>
-        </div>
-      </div>
-    </main>
-    <footer class="footer">
-      <div class="footer__container container">
-        <div class="container__github github">
-          <a class="github__link" href="https://github.com/ich-kirich" target="_blank">
-            <div class="github__pict author1">&nbsp;</div>
-          </a>
-          <a class="github__link" href="https://github.com/sqwezzy33333" target="_blank">
-            <div class="github__pict author2">&nbsp;</div>
-          </a>
-        </div>
-        <div class="container__year year">
-          <p class="year__caption">2022</p>
-        </div>
-        <div class="container__course course">
-          <a
-            class="course__link"
-            href="https://github.com/rolling-scopes-school/tasks/tree/master/stage2"
-            target="_blank"
-          >
-            <div class="course__pict">&nbsp;</div>
-          </a>
-        </div>
-      </div>
-    </footer>
-  </body>
+  <body></body>
 </html>
