+++ src/libs/index.scss
@@ -1,5 +1,5 @@
 @import 'normalize.css';
-@import '../style/filters.scss';
+@import '../pages/main/style/filters.scss';
 body::-webkit-scrollbar {
   width: 10px;
   background-color: #f9f9fd;
