+++ src/pages/main/components/draw/product/product.ts
