+++ src/pages/main/components/loader/loader.ts
