+++ src/pages/main/components/sort/sort.ts
@@ -1,17 +1,13 @@
 import { allFilters, syncURL } from '../forQueryParam/forQueryParam';
-import { IProduct } from '../../types/types';
+import { IProduct } from '../../../types/types';
 export class Sort {
-    sortBlock: HTMLSelectElement;
-
-    constructor(){
-        this.sortBlock = document.querySelector('.catalog__sort') as HTMLSelectElement;
-    }
 
     async addSortEventListeners(): Promise<void>{
+        const sortBlock = document.querySelector('.catalog__sort') as HTMLElement;
         const chooseList = document.querySelector('.sort__chooseList') as HTMLElement;
         const chooseOption = document.querySelectorAll('.chooseList__typeSort') as NodeListOf<Element>;
         this.updateNameSort();
-        this.sortBlock.addEventListener('click', (event) => {
+        sortBlock.addEventListener('click', (event) => {
             chooseList.style.display = 'block';
         });
         document.addEventListener('click', function(e) {
