+++ src/pages/main/style/filters.scss
@@ -1,4 +1,3 @@
-
 .filters {
   padding: 0px 2px 0px 4px;
   font-family: 'DM Sans', sans-serif;
