+++ src/pages/main/style/range.scss
